postman work success full

+++ src/utils/cloudinary.js
@@ -1,31 +1,34 @@
-import {v2 as cloudinary} from "cloudinary"
-import { response } from "express";
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-
-
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-
-const uploadOnCloudinary = async (localFilePath) =>{
+const uploadOnCloudinary = async (localFilePath) => {
   try {
-    if(!localFilePath) return null
-    //upload the file on cloudinary
-    cloudinary.uploader.upload(localFilePath, {
+    if (!localFilePath) return null;
+
+    const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto"
-    })
-    //file has been uploaded successfull
-    console.log("file is uploaded on cloduinary", response.url);
+    });
+
+    // console.log("File uploaded on Cloudinary:", response.url);
+    fs.unlinkSync(localFilePath)
+
     return response;
+
   } catch (error) {
-    fs.unlinkSync(localFilePath)   // remove the locally saved temporary file as the upload operation got faield
+    console.log("Cloudinary upload error:", error);
+
+    if (localFilePath && fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
     return null;
   }
-}
+};
 
-export {uploadOnCloudinary}
-	
+export { uploadOnCloudinary };
